requirements.txt
dockercloud-haproxy==1.6.7
gevent==1.1.1